requirements.txt
# === Core ===
discord.py
google-genai
anthropic
python-dotenv
pydantic
pytz

# === Database ===
pymongo
certifi

# === Web/Scraping ===
requests
beautifulsoup4
ddgs

# === Media & Voice ===
elevenlabs
openai
PyNaCl

# === Document Processing ===
pypdf
python-docx
fpdf2

# === Video ===
youtube-transcript-api

# === Utilities ===
dateparser

# === Social Media ===
tweepy

# === Messaging ===
africastalking
